+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# Version 0.12.0
+
+- Updated the Resfinder database to use the 2025-05-04 / 2.6.0 release. [PR #232](https://github.com/phac-nml/staramr/pull/232)
+- Updated the Plasmidfinder database to use the 2025-09-10 / 2.2.0 release. [PR #232](https://github.com/phac-nml/staramr/pull/232)
+- Updated StarAMR to use the `--full` option of MLST. [PR #233](https://github.com/phac-nml/staramr/pull/233)
+- StarAMR now only supports MLST v2.32.2 and greater. [PR #233](https://github.com/phac-nml/staramr/pull/233)
+- StarAMR now supports BioPython v1.86 and greater. [PR #234](https://github.com/phac-nml/staramr/pull/234)
+- Fixed many instances of warnings from BioPython and Pandas. [PR #234](https://github.com/phac-nml/staramr/pull/234)
+- Updated the gene to drug mappings. [PR #236](https://github.com/phac-nml/staramr/pull/236)
+
 # Version 0.11.1
 
 - Remove `parse_seqids` from `makeblastdb` command. Fixes issue with contig headers greater than 50 characters. [PR #225](https://github.com/phac-nml/staramr/pull/225)